+++ src/main/java/io/github/sakurawald/module/mixin/gameplay/carpet/fake_player_manager/PlayerCommandMixin.java
@@ -22,7 +22,8 @@ public abstract class PlayerCommandMixin {
 
     @Unique
     private static @NotNull String transformFakePlayerName(@NotNull String fakePlayerName) {
-        return FakePlayerManagerInitializer.config.model().transform_name.replace("%name%", fakePlayerName);
+        return FakePlayerManagerInitializer.config.model().transform_name
+            .replace("%name%", fakePlayerName);
     }
 
     @Redirect(method = "cantSpawn", at = @At(